Polish per code-quality review: dedup print controls, split venues pact
Acting on the code-quality review (no bugs — quality/cohesion):
- De-duplicate the two ~90% copy-paste scope controls in print-controls.html
  into a single _print-scope-control.html, included once per material with
  url_name/label/aria params. Rendered output is unchanged.
- Move the venue read-side contracts (VenuesServiceProtocol, VenueWithAreasDTO,
  AreaRefDTO, PrintScopeDTO) out of pacts/printing.py into pacts/venues.py —
  they're a venues concern, not a printing one.
- Drop the dead Count("spaces") annotation from AreaRepository.list_by_event
  (the consumer never reads spaces_count; it defaults to 0).
- Inline the single-use _ordered_spaces into _scoped_spaces.
