+++ src/ludamus/templates/panel/parts/print-controls.html
@@ -1,53 +1,7 @@
 {% load i18n %}
-<div class="flex items-center gap-1">
-    <a href="{% url 'panel:timetable-print' slug=current_event.slug %}"
-       target="_blank"
-       rel="noopener"
-       class="text-sm px-3 py-1.5 border border-border rounded-lg bg-bg-secondary text-foreground hover:bg-bg-tertiary transition">
-        {% translate "Print timetable" %}
-    </a>
-    {% if print_venues %}
-        <select onchange="if(this.value){window.open(this.value,'_blank','noopener');this.selectedIndex=0;}"
-                aria-label="{% translate "Print timetable for a venue or area" %}"
-                class="text-sm border border-border rounded-lg px-2 py-1.5 bg-bg-secondary text-foreground focus:outline-none focus:ring-2 focus:ring-primary">
-            <option value="">{% translate "Venue / area…" %}</option>
-            {% for v in print_venues %}
-                <optgroup label="{{ v.name }}">
-                    <option value="{% url 'panel:timetable-print' slug=current_event.slug %}?venue={{ v.slug }}">{{ v.name }}</option>
-                    {% for a in v.areas %}
-                        <option value="{% url 'panel:timetable-print' slug=current_event.slug %}?venue={{ v.slug }}&amp;area={{ a.slug }}">
-                            — {{ a.name }}
-                        </option>
-                    {% endfor %}
-                </optgroup>
-            {% endfor %}
-        </select>
-    {% endif %}
-</div>
-<div class="flex items-center gap-1">
-    <a href="{% url 'panel:timetable-print-door-cards' slug=current_event.slug %}"
-       target="_blank"
-       rel="noopener"
-       class="text-sm px-3 py-1.5 border border-border rounded-lg bg-bg-secondary text-foreground hover:bg-bg-tertiary transition">
-        {% translate "Print door cards" %}
-    </a>
-    {% if print_venues %}
-        <select onchange="if(this.value){window.open(this.value,'_blank','noopener');this.selectedIndex=0;}"
-                aria-label="{% translate "Print door cards for a venue or area" %}"
-                class="text-sm border border-border rounded-lg px-2 py-1.5 bg-bg-secondary text-foreground focus:outline-none focus:ring-2 focus:ring-primary">
-            <option value="">{% translate "Venue / area…" %}</option>
-            {% for v in print_venues %}
-                <optgroup label="{{ v.name }}">
-                    <option value="{% url 'panel:timetable-print-door-cards' slug=current_event.slug %}?venue={{ v.slug }}">
-                        {{ v.name }}
-                    </option>
-                    {% for a in v.areas %}
-                        <option value="{% url 'panel:timetable-print-door-cards' slug=current_event.slug %}?venue={{ v.slug }}&amp;area={{ a.slug }}">
-                            — {{ a.name }}
-                        </option>
-                    {% endfor %}
-                </optgroup>
-            {% endfor %}
-        </select>
-    {% endif %}
-</div>
+{% translate "Print timetable" as label_timetable %}
+{% translate "Print timetable for a venue or area" as aria_timetable %}
+{% include "panel/parts/_print-scope-control.html" with url_name="panel:timetable-print" label=label_timetable aria=aria_timetable %}
+{% translate "Print door cards" as label_door_cards %}
+{% translate "Print door cards for a venue or area" as aria_door_cards %}
+{% include "panel/parts/_print-scope-control.html" with url_name="panel:timetable-print-door-cards" label=label_door_cards aria=aria_door_cards %}
